requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
requests==2.31.0
pydantic==1.10.12
python-dotenv==0.21.1
streamlit==1.23.1
plotly==5.15.0
matplotlib==3.5.3
pandas==1.3.5